Move WatchItemPolicy to its own header. Add path type enum.
